<new>: Added connection and query to output_db

+++ schema.sql
@@ -3,7 +3,7 @@ CREATE DATABASE output_db;
 CREATE TABLE output_data (
     premise_id INT,
     data_date VARCHAR(50),
-    weather_data INT,
-    power_data INT,
+    weather_data FLOAT,
+    power_data FLOAT,
     PRIMARY KEY (premise_id, data_date)
 );
